fix logout button
