+++ src/lib/components/login.svelte
@@ -6,7 +6,7 @@
   <div class="card lg:w-1/3">
     <div class="flex min-h-full flex-col justify-center px-6 py-8">
       <h3 class="h3 mb-5">Login</h3>
-      <form class="space-y-5" method="post" action="?/login" use:enhance>
+      <form class="space-y-5" method="post" use:enhance>
         <label class="label"
           ><span>Username</span>
           <input
