--- src/routes/(authed)/account/+page.svelte
@@ -1,3 +1,0 @@
-<h1>account</h1>
-
-<p>this is the account page</p>
