+++ src/routes/(authed)/user/+page.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import type { PageData } from "./$types";
+  export let data: PageData;
+</script>
+
+
+<h2 class="h2 text-center mt-10">Welcome, {data.name}</h2>
