+++ src/routes/+layout.svelte
@@ -19,6 +19,7 @@
   } from "@floating-ui/dom";
 
   import "../app.pcss";
+  import { enhance } from "$app/forms";
 
   //requirement for popup
   storePopup.set({ computePosition, autoUpdate, offset, shift, flip, arrow });
@@ -42,7 +43,9 @@
         {#if currentPath !== "/user"}
           <a href="/login" class="btn hover:variant-soft-primary">Login</a>
         {:else}
-          <button class="btn hover:variant-soft-primary">Logout</button>
+          <form method="post" action="/logout" use:enhance>
+            <button class="btn hover:variant-soft-primary">Logout</button>
+          </form>
         {/if}
         <LightSwitch />
       </svelte:fragment>
