--- src/routes/user/+page.server.ts
@@ -1,28 +1,0 @@
-import { lucia } from "$lib/server/auth";
-import { fail, redirect } from "@sveltejs/kit";
-import type { Actions, PageServerLoad } from "./$types";
-
-export const load: PageServerLoad = async (event) => {
-  if (!event.locals.user) {
-    const srcUrl = event.url.pathname + event.url.search;
-    redirect(302, `/login?redirectTo=${srcUrl}`);
-  }
-
-  console.log(event.url);
-  return { name: event.locals.user.firstName };
-};
-
-export const actions: Actions = {
-  default: async (event) => {
-    if (!event.locals.session) {
-      return fail(401);
-    }
-    await lucia.invalidateSession(event.locals.session.id);
-    const sessionCookie = lucia.createBlankSessionCookie();
-    event.cookies.set(sessionCookie.name, sessionCookie.value, {
-      path: ".",
-      ...sessionCookie.attributes,
-    });
-    redirect(302, "/login");
-  },
-};
